review cod

+++ src/main/kotlin/example/com/firebase/firestore/client/ClientDTO.kt
@@ -1,11 +1,10 @@
 package example.com.firebase.firestore.client
 
-import com.google.cloud.firestore.annotation.DocumentId
 import kotlinx.serialization.Serializable
 
 @Serializable
 data class ClientDTO (
-    val id: Int,
+    val id: Long,
     val name: String,
     val phone: String,
     val userName: String?,
